Recommender, RecipePointDTO and sortings

+++ app/src/main/java/com/yarenbergi/yemekbul/recommender/IngredientCsv.kt
@@ -3,5 +3,5 @@ package com.yarenbergi.yemekbul.recommender
 import java.math.BigDecimal
 
 data class IngredientCsv(val ingredientName: String,
-                         val id: BigDecimal,
+                         val id: Int,
                          var point: Int)
